Transforms all activities to fragment
    * Big refactoring adding new fragments
    * Renames some views and classes
    * Removes unnecesary views file due to refactor xmls
