+++ Volontulo/app/src/main/java/com/stxnext/volontulo/ui/login/LoginFragment.java
@@ -1,0 +1,36 @@
+package com.stxnext.volontulo.ui.login;
+
+import android.content.Intent;
+import android.widget.EditText;
+import android.widget.Toast;
+
+import com.stxnext.volontulo.R;
+import com.stxnext.volontulo.VolontuloBaseFragment;
+import com.stxnext.volontulo.ui.MainHostActivity;
+
+import butterknife.Bind;
+import butterknife.OnClick;
+
+public class LoginFragment extends VolontuloBaseFragment {
+    @Bind(R.id.edit_text_email)
+    EditText editTextEmail;
+
+    @Bind(R.id.edit_text_password)
+    EditText editTextPassword;
+
+    @Override
+    protected int fragmentLayoutResource() {
+        return R.layout.fragment_login;
+    }
+
+    @OnClick(R.id.button_login)
+    public void doLogin() {
+        if (editTextEmail.getText().toString().equals("test") && editTextPassword.getText().toString().equals("test")) {
+            Intent intent = new Intent(getActivity(), MainHostActivity.class);
+            startActivity(intent);
+            getActivity().finish();
+        } else {
+            Toast.makeText(getActivity(), R.string.error_wrong_email_or_password, Toast.LENGTH_SHORT).show();
+        }
+    }
+}
