+++ Volontulo/app/src/main/java/com/stxnext/volontulo/ui/offers/OfferDetailsActivity.java
@@ -1,23 +1,24 @@
 package com.stxnext.volontulo.ui.offers;
 
 import android.os.Bundle;
-import android.support.design.widget.Snackbar;
+import android.support.v4.app.FragmentManager;
 import android.support.v4.app.NavUtils;
 import android.view.MenuItem;
-import android.view.View;
 
 import com.stxnext.volontulo.R;
 import com.stxnext.volontulo.VolontuloBaseActivity;
-
-import butterknife.OnClick;
 
 public class OfferDetailsActivity extends VolontuloBaseActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_offers_details);
+        setContentView(R.layout.activity_nodrawer);
         init(R.string.title_activity_offers_details);
+        final FragmentManager fragmentManager = getSupportFragmentManager();
+        fragmentManager.beginTransaction()
+            .replace(R.id.content, new OfferDetailsFragment())
+            .commit();
     }
 
     @Override
@@ -28,11 +29,6 @@ public class OfferDetailsActivity extends VolontuloBaseActivity {
                 return true;
         }
         return super.onOptionsItemSelected(item);
-    }
-
-    @OnClick(R.id.button_step_out)
-    public void doStepOut(View view) {
-        Snackbar.make(view, "Zgłosiłeś się!!!", Snackbar.LENGTH_SHORT).show();
     }
 
 }
