+++ Volontulo/app/src/main/res/layout/fragment_offer_add.xml
@@ -6,12 +6,6 @@
     android:layout_height="match_parent"
     android:orientation="vertical">
 
-    <include
-        android:id="@+id/appbar"
-        layout="@layout/app_bar"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content" />
-
     <ScrollView
         android:id="@+id/scroller"
         android:layout_width="match_parent"
