+++ Volontulo/app/src/main/res/layout/fragment_offers_details.xml
@@ -7,9 +7,8 @@
     android:orientation="vertical"
     tools:context=".ui.offers.OfferDetailsActivity">
 
-    <include layout="@layout/app_bar" />
-
-    <include layout="@layout/view_offer_details"
+    <include
+        layout="@layout/view_offer_details"
         android:layout_width="match_parent"
         android:layout_height="0dp"
         android:layout_weight="1" />
@@ -17,10 +16,10 @@
     <LinearLayout
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:background="#DA4453"
-        android:orientation="vertical"
         android:layout_gravity="bottom|end"
+        android:background="#DA4453"
         android:gravity="center"
+        android:orientation="vertical"
         android:visibility="visible">
 
         <TextView
@@ -40,7 +39,7 @@
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:text="Zgłoś się na ten wolontariat"
-            tools:text="Zgłoś się na ten wolontariat"/>
+            tools:text="Zgłoś się na ten wolontariat" />
 
     </LinearLayout>
