+++ Volontulo/app/src/main/res/layout/view_offer_details.xml
@@ -10,7 +10,7 @@
     android:paddingTop="@dimen/activity_vertical_margin"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
     tools:context=".ui.offers.OfferDetailsActivity"
-    tools:showIn="@layout/activity_offers_details">
+    tools:showIn="@layout/fragment_offers_details">
 
     <LinearLayout
         android:id="@+id/header"
